Module empty_template::lib::neopixel

Welcome to the neopixel library source code!

By: Jonathan Zurita

Date: 11/18/2024

Version: 0.1.0

What is the function of this library?

This library allows the user to call user-friendly functions in order to manipulate WS2812 neopixels.

IMPORTANT NOTES

Please edit the pub const variables as needed

(e.g. change NUM_LEDS to 24 instead of default value of 8.)

Why do use a watchdog timer as an argument?

I decided to include a watchdog timer in order to protect against possible lockups in DMA transmission. Thankfully utilizing a WDT requires minimal overhead. :)

Release Notes

V0.1.0 - Released neopixel library to the public

Constants

ARR_SIZE

ARR_SIZE = NUM_LEDS * PACKET_SIZE

NEOPIXEL_LOGIC_0

SPI generate waveforms (LOGIC_1)

NEOPIXEL_LOGIC_1

SPI generate waveforms (LOGIC_0)

NUM_LEDS

Desired number of neopixels to control

PACKET_SIZE

Number of color bits per neopixel

TURN_OFF ๐

Essentially hex color code for black (no light)

Functions

neopixel_init_buffer

<<< Click on me to see function desc!

neopixel_set_data

<<< Click on me to see function desc!

neopixel_set_entire_data

<<< Click on me to see function desc!

spi_dma_send

spi_dma_send_breathing